jarray {rJava}	R Documentation

Java array handling functions

Description

.jarray takes a vector (or a list of Java references) as its argument, creates a Java array containing the elements of the vector (or list) and returns a reference to such newly created array.

.jevalArray takes a reference to a Java array and returns its contents (if possible).

Usage


.jarray(x, contents.class = NULL)
.jevalArray(obj, rawJNIRefSignature = NULL, silent = FALSE)

Arguments

x 	 vector or a list of Java references

contents.class 	 common class of the contained objects, see details

obj 	 Java object reference to an array that is to be evaluated

rawJNIRefSignature 	 JNI signature that whould be used for conversion. If set to NULL, the signature is detected automatically.

silent 	 if set to true, warnings are suppressed

Details

.jarray: The input can be either a vector of some sort (such as numeric, integer, logical, ...) or a list of Java references. The contents is pushed to the Java side and a corresponding array is created. The type of the array depends on the input vector type. For example numeric vector creates double[] array, integer vector creates int[] array, character vector String[] array and so on. If x is a list, it must contain Java references only (or NULLs which will be treaded as NULL references).

The contents.class parameter is used only if x is a list of Java object references and it can specify the class that will be used for all objects in the array. If set to NULL no assumption is made and java/lang/Object will be used. Use with care and only if you know what you're doing - you can always use .jcast to cast the entire array to another type even if you use a more general object type.

The result is a reference to the newly created array.

The inverse function which fetches the elements of an array reference is .jevalArray.

.jevalArray currently supports only a subset of all possible array types. Recursive arrays are handled by returning a list of references which can then be evaluated separately.

Value

.jarray returns a Java array reference (jarrayRef) to an array created with the supplied contents.
.jevalArray returns the contents of the array object.

Examples


## Not run: 
a <- .jarray(1:10)
print(a)
.jevalArray(a)
b <- .jarray(c("hello","world"))
print(b)
c <- .jarray(list(a,b))
print(c)
## End(Not run)

[Package rJava version 0.4-12 Index]