jarrayRef-class {rJava}	R Documentation

Class "jarrayRef" Reference to an array Java object

Description

This class is a subclass of jobjRef-class and represents a reference to an array Java object.

Objects from the Class

Objects cannot be created directly, but only as the retun value of .jcall function.

Slots

jsig:

JNI signature of the array type

jobj:

Internal identifier of the object

jclass:

Inherited from jobjRef, but unspeficied

Extends

Class "jobjRef", directly.

Methods

No methods defined with class "jarrayRef" in the signature.

Author(s)

Simon Urbanek

See Also

.jcall or jobjRef-class

[Package rJava version 0.4-12 Index]