jcall {rJava}	R Documentation

Call a Java method

Description

.jcall calls a Java method with the supplied arguments.

Usage


.jcall(obj, returnSig = "V", method, ..., evalArray = TRUE, 
    evalString = TRUE, check = TRUE, interface = "RcallMethod")

Arguments

obj 	 Java object (jobjRef as returned by .jcall or .jnew) or fully qualified class name in JNI notation (e.g. "java/lang/String").

returnSig 	 Return signature in JNI notation (e.g. "V" for void, "[I" for int[] etc.). For convenience additional type "S" is supported and expanded to "Ljava/lang/String;", re-mapping "T" to represent the type short.

method 	 The name of the method to be called

... 	 Any parametes that will be passed to the Java method. The parameter types are determined automatically and/or taken from the jobjRef object. All named parameters are discarded.

evalArray 	 This flag determines whether the array return value is evaluated (TRUE) or passed back as Java object reference (FALSE).

evalString 	 This flag determines whether string result is returned as characters or as Java object reference.

check 	 If set to TRUE then checks for exceptions are performed before and after the call using .jcheck(silent=FALSE). This is usually the desired behavior, because all calls fail until an expection is cleared.

interface 	 This option is experimental and specifies the interface used for calling the Java method; the current implementation supports two interfaces:

"RcallMethod"

the default interface.

"RcallSyncMethod"

synchronized call of a method. This has simmilar effect as using synchronize in Java.

Details

.jcall requires exact match of argument and return types. For higher efficiency .jcall doens't perform any lookup in the reflection tables. This means that passing subclasses of the classes present in the method definition requires explicit casting using .jcast. Passing null arguments also needs a proper class specification with .jnull.

Java types long and float have no corresponding types in R and therefore any such parameters must be flagged as such using .jfloat and .jlong functions respectively.

Java also distinguishes scalar and array types whereas R doesn't have the concept of a scalar. In R a scalar is basically a vector (called array in Java-speak) of the length 1. Therefore passing vectors of the length 1 is ambiguous. .jcall assumes that any vector of the length 1 that corresponds to a native Java type is a scalar. All other vectors are passed as arrays. Therefore it is important to use .jarray if an arbitrary vector (including those of the length 1) is to be passed as an array parameter.

Value

Returns the result of the method.

See Also

.jnew, .jcast, .jnull, .jarray

Examples


## Not run: 
.jcall("java/lang/System","S","getProperty","os.name")
f <- .jnew("java/awt/Frame","Hello")
.jcall(f,,"setVisible",TRUE)
## End(Not run)

[Package rJava version 0.4-12 Index]