jclassName-class {rJava}	R Documentation

Class "jclassName" - a representation of a Java class name

Description

This class holds a name of a class in Java.

Objects from the Class

Objects of this class should *not* be created directly. Instead, the function J should be used to create new objects of this class.

Slots

name:

Name of the class (in source code notation)

jobj:

Object representing the class in Java

Methods

The objects of class jclassName are used indirectly to be able to create new Java objects via new such as new(J("java.lang.String"), "foo") or to use the $ convenience operator on static classes, such as J("java.lang.Double")$parseDouble("10.2").

as.character

signature(x = "jclassName"): returns the class name as a string vector of length one.

Author(s)

Simon Urbanek

See Also

J, new

[Package rJava version 0.9-13 Index]