jengine {rJava}	R Documentation

Java callback engineCast a Java object to another class

Description

.jengine obtains the current callback engine or starts it.

Usage

.jengine(start=FALSE, silent=FALSE)

Arguments

start
if set to TRUE then the callback engine is started if it is not yet active

silent
if set to TRUE then NULL is returned if there is no engine available. Otherwise an error is raised

Details

.jengine can be used to detect whether the engine was started or to start the engine.

Before any callbacks from Java into R can be performed, the Java callback engine must be initialized, loading Java/R Interface (JRI). If JRI was not started and start is set to TRUE then .jengine will load necessary classes and start it.

Note that JRI is an optional part of rJava and requires R shared library at the moment. By default rJava will continue with installation even if JRI cannot be built.

Value

Returns a Java object reference (jobjRef) to the current Java callback engine.

See Also

.jcall

Examples

## Not run: 
.jengine(TRUE)

## End(Not run)

[Package rJava version 0.9-13 Index]