jfloat-class {rJava}	R Documentation

Classes "jfloat", "jlong", "jbyte" and "jchar" specify Java native types that are not native in R

Description

These classes wrap a numeric vector to be treated as float or long argument when passed to Java and an integer vector to be treated as byte or char. R doesn't distinguish between double and float, but Java does. In order to satisfy object types, numeric vectors that should be converted to floats or long on the Java side must be wrapped in this class. In addition jbyte must be used when passing scalar byte (but not byte arrays, those are mapped into RAW vectors). Finally jchar it used when mapping integer vectors into unicode Java character vectors.

Objects from the Class

Objects can be created by calling .jfloat, .jlong, .jbyte or .jchar respectively.

Slots

.Data:

Payload

Extends

"jfloat" and "jlong": Class "numeric", from data part. Class "vector", by class "numeric".

"jbyte" and "jchar": Class "integer", from data part. Class "vector", by class "integer".

Methods

"jfloat" and "jlong" have no methods other than those inherited from "numeric". "jbyte" and "jchar" have no methods other than those inherited from "integer".

Author(s)

Simon Urbanek

See Also

.jfloat, .jlong, .jbyte, .jchar and .jcall

[Package rJava version 0.4-12 Index]