.jgc {rJava}	R Documentation

Invoke Java Garbage Collection

Description

.jgc invokes the R and Java garbage collectors.

Usage

.jgc(R.gc = TRUE, ...)

Arguments

R.gc
logical, if TRUE then gc(...) is called first, if FALSE only Java garbage collector is called

...
any additional parameters passed to gc()

Details

.jgc invokes the R garbage collector (unless R.gc=FALSE) which removes any unused Java references and then invokes the Java garbage collector to reclaim Java heap space.

Author(s)

Simon Urbanek

[Package rJava version 0.9-13 Index]