show {rJava}	R Documentation

Show a Java Object Reference

Description

Display a Java object reference in a descriptive, textual form. The default implementation calls toString Java method to obtain object's printable value and uses calls show on the resulting string garnished with additional details.

Methods

show

signature(object = "jobjRef"): ...

show

signature(object = "jarrayRef"): ...

[Package rJava version 0.4-12 Index]