jarrayRef-class {rJava}	R Documentation

Class "jarrayRef" Reference to an array Java object

Description

This class is a subclass of jobjRef-class and represents a reference to an array Java object.

Objects from the Class

Objects cannot be created directly, but only as the return value of .jcall function.

Slots

jsig:

JNI signature of the array type

jobj:

Internal identifier of the object

jclass:

Inherited from jobjRef, but unspecified

Methods

[

signature(x = "jarrayRef"): not yet implemented

[[

signature(x = "jarrayRef"): R indexing of java arrays

[[<-

signature(x = "jarrayRef"): replacement method

head

signature(x = "jarrayRef"): head of the java array

tail

signature(x = "jarrayRef"): tail of the java array

length

signature(object = "jarrayRef"): Number of java objects in the java array

str

signature(object = "jarrayRef"): ...

unique

signature(x = "jarrayRef"): not yet implemented

duplicated

signature(x = "jarrayRef"): not yet implemented

anyDuplicated

signature(x = "jarrayRef"): not yet implemented

sort

signature(x = "jarrayRef"): not yet implemented

rev

signature(x = "jarrayRef"): not yet implemented

min

signature(x = "jarrayRef"): not yet implemented

max

signature(x = "jarrayRef"): not yet implemented

range

signature(x = "jarrayRef"): not yet implemented

Extends

Class "jobjRef", directly.

Author(s)

Simon Urbanek

See Also

.jcall or jobjRef jrectRef for rectangular arrays

[Package rJava version 0.9-13 Index]