jobjRef-class {rJava}	R Documentation

Class "jobjRef" - Reference to a Java object

Description

This class describes a reference to an object held in a JavaVM.

Objects from the Class

Objects of this class should *not* be created directly. Instead, the function .jnew should be use to create new Java objects. They can also be created as results of the .jcall function.

Slots

jobj:

Internal identifier of the object (external pointer to be precise)

jclass:

Java class name of the object (in JNI notation)

Java-side attributes are not accessed via slots, but the $ operator instead.

Methods

This object's Java methods are not accessed directly. Instead, .jcall JNI-API should be used for invoking Java methods. For convenience the $ operator can be used to call methods via reflection API.

Author(s)

Simon Urbanek

See Also

.jnew, .jcall or jarrayRef-class

[Package rJava version 0.9-13 Index]