loader {rJava}	R Documentation

Java class loader

Description

.jaddClassPath adds directories or JAR files to the class path.

.jclassPath returns a vector containg the current entries in the class path

Usage

.jaddClassPath(path)
.jclassPath()

Arguments

path
character string vector listing the paths to add to the class path

Value

.jclassPath returns a character vector listing the class path sequence.

Examples

## Not run: 
.jaddClassPath("/my/jars/foo.jar","/my/classes/")
print(.jclassPath())

## End(Not run)

[Package rJava version 0.9-13 Index]