show {rJava}	R Documentation

Show a Java Object Reference

Description

Display a Java object reference in a descriptive, textual form. The default implementation calls toString Java method to obtain object's printable value and uses calls show on the resulting string garnished with additional details.

Methods

show

signature(object = "jobjRef"): ...

show

signature(object = "jarrayRef"): ...

show

signature(object = "jclassName"): ...

str

signature(object = "jobjRef"): currently identical to show

[Package rJava version 0.9-13 Index]